refactor: extract path resolution from Frontend.Manager to Resolve module
System.cmd calls (xcodebuild, ps) for renderer and tty path detection
were running inside the GenServer init, blocking the editor process.
Move all path resolution to MingaEditor.Frontend.Resolve and call it
from the supervisor before building the child spec. The GenServer init
is now a pure state constructor with no blocking I/O.
